+++ protocol-ops/src/upgrade_verification/versions/v31/elements/mod.rs
@@ -4,4 +4,5 @@ pub mod fixed_force_deployment;
 pub mod governance_stage_calls;
 pub mod initialize_data_new_chain;
 pub mod protocol_version;
+pub mod rpc_state;
 pub mod set_new_version_upgrade;
